👌 IMPROVE: Make gitconfig single source of truth for emoji log
This commit deletes the fish functions and updates the `.gitconfig`. I've replaced the functions with fish abbreviations which rely on the `.gitconfig`.

The git aliases are now the single source of truth for the emoji log git commands.

I've also removed the 'git commit & git push in one command' (`gcap`), because I've rather have the ability to push manually.